Added expiration check

+++ setpass.py
@@ -1,3 +1,4 @@
+import datetime
 import json
 import uuid
 
@@ -14,7 +15,14 @@ app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///'
 db = SQLAlchemy(app)
 
 
+EXPIRES_AFTER_SECONDS = 24 * 3600
+
+
 class TokenNotFoundException(Exception):
+    pass
+
+
+class TokenExpiredException(Exception):
     pass
 
 
@@ -26,20 +34,25 @@ class User(db.Model):
     __tablename__ = 'users'
     id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.String(64), unique=True)
-    token = db.Column(db.String(64), nullable=False)
+    token = db.Column(db.String(64), unique=True)
     password = db.Column(db.String(64), nullable=False)
+    updated_at = db.Column(db.DateTime, nullable=False)
 
     def __init__(self, user_id, token, password):
         self.user_id = user_id
         self.token = token
         self.password = password
+        self.update_timestamp()
+
+    def update_timestamp(self):
+        self.updated_at = datetime.datetime.utcnow()
 
     def __repr__(self):
         return '<User %r, Token %r>' % (self.user_id, self.token)
 
     @staticmethod
-    def find(token):
-        return User.query.filter_by(token=token).first()
+    def find(**kwargs):
+        return User.query.filter_by(**kwargs).first()
 
 
 db.create_all()
@@ -62,6 +75,8 @@ def set_password():
         _set_password(token, password)
     except TokenNotFoundException:
         return Response(response='Token not found', status=404)
+    except TokenExpiredException:
+        return Response(response='Token expired', status=403)
 
     return Response(status=200)
 
@@ -80,12 +95,16 @@ def _set_openstack_password(user_id, old_password, new_password):
 
 def _set_password(token, password):
     # Find user for token
-    user = User.find(token)
+    user = User.find(token=token)
 
     if user is None:
         raise TokenNotFoundException
 
-    _set_openstack_password(user.user_id, user.password, password)
+    delta = datetime.datetime.utcnow() - user.updated_at
+    if delta.total_seconds() > EXPIRES_AFTER_SECONDS:
+        raise TokenExpiredException
+
+    #_set_openstack_password(user.user_id, user.password, password)
 
     db.session.delete(user)
     db.session.commit()
@@ -95,15 +114,20 @@ def _set_password(token, password):
 def add(user_id):
     password = request.data
 
-    user = User(
-        user_id=user_id,
-        token=str(uuid.uuid4()),
-        password=password
-    )
+    user = User.find(user_id=user_id)
+    if user:
+        user.password = password
+        user.token = str(uuid.uuid4())
+        user.update_timestamp()
+    else:
+        user = User(
+            user_id=user_id,
+            token=str(uuid.uuid4()),
+            password=password
+        )
+        db.session.add(user)
 
-    db.session.add(user)
     db.session.commit()
-
     return Response(response=user.token, status=200)
